refactor(nova): own the per-instance metrics cron on compute nodes
Review feedback, and the better placement. The cron runs hex_sdk watcher_instance_metrics, which reads the local libvirt domains -- so it belongs to the module that runs where the domains are. config_nova.cpp gates it on IsCompute, exactly as nova's own compute services are gated a few lines below.

It sat in config_prometheus.cpp for a weaker reason: config_watcher.cpp returns early unless IsControl, while the instances live on compute, and config_prometheus.cpp's Commit happened to run on every role. That made it work rather than made it right, and it wrote the cron on control-only and storage nodes too, where the generator has no virsh and produces an empty file. Now those nodes get no cron at all, and the writer unlinks it if the role or the service is disabled. config_prometheus.cpp keeps only what is genuinely its own: node_exporter's --collector.textfile.directory flag and the scrape.

Verified on accept-3cc and jim-1cc. All four nodes are control-converged, so IsCompute matches and each gets the cron; the node hosting the instance emits both series and the rest emit none, which is the intended no-op. config_nova.o carries the cron literals and config_prometheus.o no longer references them. Re-running the three verified strategies afterwards gives host_ram_usage_percent 0.0873 and host_cpu_usage_percent 0.0244 on the instance's node, unchanged. Also closes a gap: jim-1cc had the datasource half deployed but never the collector, so it now matches accept-3cc.

Not exercised: the control-only path where the cron should be absent. Every node on both lab clusters is control-converged, so no topology here can reach it; the unlink branch is reasoned from the same IsCompute gate nova already relies on.
